Test: `splitToNumberArray` 테스트 작성
